+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 All notable changes to this project will be documented in this file.  The
 format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
-#### Added
+## [1.5] (Android versionCode 11) - 2019-07-30
 #### Changed
 - [Issue #18](https://github.com/kuhrusty/z15/issues/18): switch from
   `Resources.getDrawable()` to `AppCompatResources.getDrawable()` and